Update .gitignore: Ignore target and all git repos
`target` is a Rust build artifact and must be ignored. The Git repos
that are referenced in `build.rs` should also be ignored. I am
considering moving them into `target` as well, but I don't think that is
where they are meant to be.

Contact: Sam Christy <samchristywork@gmail.com>